requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0

# FAISS for similarity search
faiss-cpu>=1.7.4

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Utilities
pathlib>=1.0.1
pickle5>=0.0.11

# Optional: For GPU support
# faiss-gpu>=1.7.4
# torch with CUDA support

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
